Create extended_genes.gff
Ive extended the gene models in the TGAC gene predictions. I did the following

Align the RNAseq reads from KW1 against the KW1 assembly, using BWA
For each `gene` model in the TGAC gene predictions, that was within 100nt of another gene I extracted reads on the same strand that fell within -1000nt  of the start or 1000nt of the end
With these reads, starting with the start and end of the gene I followed read overlaps as far as possible, until reads no longer overlapped. The most distal read then counted as the new gene start/end.
The below GFF shows all extended gene features.